requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
humanize==0.5.1
Pillow
pytz==2019.2
